requirements.txt
gymnasium
playwright
google-generativeai
numpy
Pillow
Markdown
beautifulsoup4
axe-core-python
tqdm

